Do not insert again in the database unchanged users and contacts
The function TelegramQml::contactsGetContacts_slt is invoked by signal
contactsGetContactsAnswer which is emitted by libqtelegram-ae.

Here the signal is emitted by Telegram::onContactsContacts and
Telegram::onContactsContactsNotModified.

The former set the second parameter to true, so this patch doesn't
change its behavior, but the latter set the modified parameter to
false, because it serves contacts and users from cache.

Since they are from caches, there is no need to call insertUser and
insertContact and add them again to the database

On my account, this removes ~250 queries to the database at startup